[skip ci] Modularize SVG
