+++ plugins/svg/src/SvgFeatureRenderer/components/FeatureGlyph.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react'
 
 import FeatureLabel from './FeatureLabel'
 
-import type { DisplayModel } from './util'
+import type { DisplayModel, ViewParams } from './types'
 import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
 import type { Feature, Region } from '@jbrowse/core/util'
 import type { SceneGraph } from '@jbrowse/core/util/layouts'
@@ -25,12 +25,7 @@ const FeatureGlyph = observer(function (props: {
   reversed?: boolean
   topLevel?: boolean
   region: Region
-  viewParams: {
-    end: number
-    start: number
-    offsetPx: number
-    offsetPx1: number
-  }
+  viewParams: ViewParams
   bpPerPx: number
 }) {
   const {
