+++ plugins/svg/src/SvgFeatureRenderer/components/FeatureLabel.tsx
@@ -5,7 +5,7 @@ import { useTheme } from '@mui/material'
 import { observer } from 'mobx-react'
 import { isAlive, isStateTreeNode } from 'mobx-state-tree'
 
-import type { DisplayModel } from './util'
+import type { DisplayModel } from './types'
 import type { Feature, Region } from '@jbrowse/core/util'
 
 interface ViewParams {
