+++ plugins/svg/src/SvgFeatureRenderer/components/Subfeatures.tsx
@@ -1,9 +1,10 @@
 import { readConfObject } from '@jbrowse/core/configuration'
 import { observer } from 'mobx-react'
 
-import { chooseGlyphComponent, layOut, layOutFeature } from './util'
+import { chooseGlyphComponent } from './chooseGlyphComponent'
+import { layOut, layOutFeature } from './util'
 
-import type { ExtraGlyphValidator } from './util'
+import type { ExtraGlyphValidator } from './types'
 import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
 import type { SceneGraph } from '@jbrowse/core/util/layouts'
 import type { Feature } from '@jbrowse/core/util/simpleFeature'
