+++ plugins/svg/src/SvgFeatureRenderer/components/SvgOverlay.tsx
@@ -3,9 +3,8 @@ import { useEffect, useState } from 'react'
 import { bpSpanPx } from '@jbrowse/core/util'
 import { observer } from 'mobx-react'
 
+import type { LayoutRecord } from './types'
 import type { Feature, Region } from '@jbrowse/core/util'
-
-type LayoutRecord = [number, number, number, number]
 
 interface OverlayRectProps extends React.SVGProps<SVGRectElement> {
   rect?: LayoutRecord
